requirements.txt
django
djangorestframework
certifi==2023.7.22
charset-normalizer==3.3.0
defusedxml==0.7.1
httpie==3.2.2
idna==3.4
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
Pygments==2.16.1
PySocks==1.7.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.6.0
urllib3==2.0.6
